chore: refactor usePLaylistSettings
